新增球球大作战游戏测试工程

+++ src/github.com/KylinHe/aliensboot-server/data/atargame.go
@@ -76,7 +76,7 @@ type AtarGameBase struct {
 	MaxThornScore int32
 	SplitDuration	int32
 	GameTime int32
-	MapWidth int32
+	MapWidth float32
 	ThornColorID int32
 	ThornColor []int32
 	Colors []Color
